requirements.txt
Flask
torch
transformers
Pillow
requests
scikit-learn
